af_mydrc.c: 400->500 ms rpm filter size, different 100 ms weights
When we try to measure the loudness of a 100 ms fame by combining
4 frames, this is counter-intuitive: the middle frame isn't centered,
resulting in a shift (and hard-to-reason-about code).  It's much easier
to evaluate the loudness of a center frame by combining 5 frames.
We can further assign weights to the frames, with the middle frame
getting the highest weight.