aacgain15: added options --rg=0.5 --db=+0.0
With --rg=0.333, it is possible to reduce the fraction of replaygain
in the resulting gain calculation (by default the reulting gain is the
average of replaygain and ebur128).  With --db=+2.0, it is possible to
normalize to the target loudness -16 LUFS instead of default -18 LUFS.

I also replace POSIX::floor in igain calculation with sprintf rounding.
Not sure why I used POSIX::floor, perhaps I only considered turning
the volume up, e.g. going from -23 LUFS to -18 LUFS.  Now I also have
to consider going down from -16 LUFS to -18 LUFS.